# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured context for the error, or null.",
            "title": "Detail"
          },
          "error_code": {
            "description": "Machine-readable code. Branch on this, not on the message.",
            "enum": [
              "EMPTY_STATEMENT",
              "MULTIPLE_STATEMENTS",
              "NON_SELECT_STATEMENT",
              "FORBIDDEN_KEYWORD",
              "FORBIDDEN_FUNCTION",
              "INVALID_SQL",
              "QUERY_EXECUTION_ERROR",
              "INVALID_CHART_CONFIG",
              "INVALID_CONNECTION_CONFIG",
              "DB_TLS_REQUIRED",
              "CONNECTION_PAUSED",
              "ROW_LIMIT_EXCEEDED",
              "SQL_TOO_LONG",
              "RESULT_TOO_LARGE",
              "WEAK_PASSWORD",
              "DB_AUTH_FAILED",
              "DB_PERMISSION_DENIED",
              "INVALID_CREDENTIALS",
              "NOT_AUTHENTICATED",
              "ACCOUNT_LOCKED",
              "PASSWORD_CHANGE_REQUIRED",
              "FORBIDDEN",
              "CONNECTION_NOT_FOUND",
              "QUERY_NOT_FOUND",
              "TABLE_NOT_FOUND",
              "DASHBOARD_NOT_FOUND",
              "USER_NOT_FOUND",
              "DUPLICATE_NAME",
              "DUPLICATE_EMAIL",
              "LAST_ADMIN",
              "QUERY_FROZEN",
              "REQUEST_VALIDATION_ERROR",
              "RATE_LIMITED",
              "DB_UNREACHABLE",
              "QUERY_TIMEOUT",
              "INTERNAL_ERROR",
              "SERVICE_NOT_READY"
            ],
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "AuditEntryRead": {
        "description": "One row of the audit log, with the actor's email resolved for display.\n\nMirrors ``app.features.audit.models.AuditLog`` field for field, plus\n``actor_email`` - the log stores ``actor_id`` because accounts are never\ndeleted and an id is what survives an email changing, but a page rendering\nthe log for a human needs something readable without a second lookup per\nrow.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/AuditAction"
          },
          "actor_email": {
            "title": "Actor Email",
            "type": "string"
          },
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "actor_id",
          "actor_email",
          "action",
          "target_type",
          "target_id",
          "detail",
          "created_at"
        ],
        "title": "AuditEntryRead",
        "type": "object"
      },
...
      "UserCreateResponse": {
        "description": "The one and only place a freshly issued temporary password appears.",
        "properties": {
          "temporary_password": {
            "title": "Temporary Password",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserRead"
          }
        },
        "required": [
          "user",
          "temporary_password"
        ],
        "title": "UserCreateResponse",
        "type": "object"
      },
      "UserRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active",
          "must_change_password",
          "last_login_at",
          "created_at"
        ],
        "title": "UserRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Schema-agnostic backend for saved, read-only SQL exposed as chart-ready JSON. All fraud logic lives in the SQL you save.",
    "title": "Fraud Analyzer Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/audit-log": {
      "get": {
        "description": "The most recent audit entries, newest first, with the actor's email\nresolved.\n\nCapped at ``limit`` (default AUDIT_LOG_DEFAULT_LIMIT, maximum\nAUDIT_LOG_MAX_LIMIT) rather than returning the whole table. Audit rows\nare never deleted by design, so the table only grows for as long as the\ninstallation runs, and an unbounded ``SELECT *`` here would eventually\nhave to serialise and ship every row that was ever written. Newest first\nwith a cap, rather than any other slice, because the entries an admin\nreading this screen actually needs are the ones nearest to whatever they\nare investigating - a login lockout, a role change someone is asking\nabout - and that is always the most recent activity, not the oldest.\n\nThe table stores ``actor_id`` rather than an email because accounts are\nnever deleted and an id is what survives an email changing (see\n``AuditEntryRead``'s docstring), but a screen rendering the log for a\nhuman needs something readable without a second lookup per row - hence\nthe join here rather than in the schema or the model.",
        "operationId": "list_audit_log_audit_log_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEntryRead"
                  },
                  "title": "Response List Audit Log Audit Log Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "List Audit Log",
        "tags": [
          "users"
        ]
      }
    },
...
    "/users": {
      "get": {
        "description": "Every account, newest first, for the admin account list.",
        "operationId": "list_users_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  },
                  "title": "Response List Users Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "List Users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "description": "Open an account with a system-generated password.\n\nThe plaintext temporary password appears exactly once, in this response,\nand is never recoverable afterwards - not from a later GET on this\naccount, and not from the audit entry ``user_service.create_user`` writes\nfor it, which records only the email and role.",
        "operationId": "create_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Create User",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{user_id}": {
      "patch": {
        "description": "Deactivate/reactivate an account, change its role, or both at once.\n\nRefused with ``LAST_ADMIN`` (409) if the change would leave the\ninstallation with no active administrator - see\n``user_service.guard_last_admin``. Refused with ``USER_NOT_FOUND`` (404)\nif ``user_id`` does not name an existing account.",
        "operationId": "update_user_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Update User",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{user_id}/reset-password": {
      "post": {
        "description": "Issue a fresh temporary password and end every live session on this\naccount at once.\n\nA reset is issued exactly when an account is suspected compromised or its\nowner is locked out; either way, the person currently holding the old\ncredential must not keep running alongside the new one.",
        "operationId": "reset_password_users__user_id__reset_password_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemporaryPasswordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Reset Password",
        "tags": [
          "users"
        ]
      }
    }
  }
}